PHP Private Gallery : Code Organization

+++ project_site/private_gallery.php
@@ -33,36 +33,41 @@
 
         <div class="content">
             
-            <?php //if $_SESSION['user']->getID() == 2 ?>
-
-            <div class="edit_page">
-            <?php 
-                foreach($_SESSION['private_gal']->getPictures() as $picture) {
-                echo "<img src='private_images/".$picture->getPath()."' alt='' height=100 /> {$picture->getName()} | {$picture->getDesc()} | {$picture->getDate()} 
-                 
-                 
-                 <a href='?action=edit&id={$picture->getId()}'    class='btn btn-default' role='button'> Edit </a>
-                 <a href='?action=delete&id={$picture->getId()}'  class='btn btn-default' role='button'> Delete </a>
-                
-                
-                
-                </br></br>";
-                
-                
-                }
-                ?>
-            </div>
+            <?php
             
-            
-            <?php //else ?>
-
-            <div class="gallery">
-
-                <?php /*
-
+            // IF THE USER IS THE ADMIN => DISPLAY ADMIN EDIT PAGE
+            if( $_SESSION['user'] instanceof User && $_SESSION['user']->getID() == 2 ) {
+                
+                echo "<div class='edit_page'><br>";
+                
+                echo "Picture | Picture Title | Picture Description | Date <br>";
+                
                 foreach($_SESSION['private_gal']->getPictures() as $picture) {
                     
-                    //if($picture->getState()==2) {
+                    echo "
+                    
+                        <img src='private_images/".$picture->getPath()."' alt='".$picture->getTitle()."' height='100px' /> {$picture->getName()} | {$picture->getDesc()} | {$picture->getDate()} 
+                    
+                        <a href='?action=edit&id={$picture->getId()}'    class='btn btn-default' role='button'> Edit </a>
+                        
+                        <a href='?action=delete&id={$picture->getId()}'  class='btn btn-default' role='button'> Delete </a>
+                        
+                        <br/><br/>
+                        
+                        ";  
+                
+                }
+                
+                echo "</div>";
+                
+            }
+            
+            // ELSE (IF IT IS A NORMAL USER) => DISPLAY USER GALLERY
+            else {
+                
+                echo "<div class='gallery'>";
+
+                foreach($_SESSION['private_gal']->getPictures() as $picture) {
                         
                             echo "
                                 <div class=\"gal_element\"> 
@@ -77,12 +82,14 @@
 
                                 </div>
                             ";
-                    //}
                 }
-
-                */ ?>
-
-            </div>
+                
+                echo "</div>";
+                
+            }
+            
+            ?>
+            
         </div>
 
         <footer>
